Change all ids from int to long type

+++ app/src/main/java/com/test/smartnotes/ViewNoteActivity.java
@@ -25,7 +25,7 @@ public class ViewNoteActivity extends AppCompatActivity {
         setSupportActionBar(toolbar);
 
         Intent intent = getIntent();
-        int id = intent.getIntExtra("id", 1);
+        long id = intent.getLongExtra("id", 1);
 
         NoteData note = DBAdapter.getNoteData(id);
